requirements.txt
pandas>=2.2.0
numpy>=1.26.0
pymatgen>=2024.6.10
streamlit>=1.40.0
plotly>=5.24.0
tqdm>=4.66.0
requests>=2.31.0
anthropic>=0.49.0
mp-api>=0.45.2
matminer>=0.9.3
jarvis-tools>=2024.10.30
reportlab>=4.2.0
torch>=2.11.0
scikit-learn>=1.8.0
